pyproject.toml
[project]
name = "alfamanager-crm-backend"
version = "0.1.0"
description = "crm to alfa manager"
authors = [
    {name = "AnthonyLoche", email = "anthonylocheifc@gmail.com"},
]
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "python-dotenv>=1.1.0",
    "django-cors-headers>=4.7.0",
    "django-extensions>=4.1",
    "pydotplus>=2.0.2",
    "faker>=37.3.0",
    "djangorestframework-simplejwt>=5.5.1",
    "dj-database-url>=3.0.0",
    "psycopg2-binary>=2.9.10",
    "uvicorn>=0.35.0",
    "whitenoise[brotli]>=6.6.0",
    "gunicorn>=21.2.0",
    "django-filter>=25.1",
    "ipython>=9.7.0",
    "requests>=2.32.5",
    "rapidfuzz>=3.14.5",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"


[tool.pdm]
distribution = true


[tool.pdm.scripts]
dev = "python src/manage.py runserver"
devc = "bash -c 'clear && pdm run python src/manage.py runserver'"
devweb = "pdm run uvicorn config.asgi:application --reload --port 8000"
runserver = "python src/manage.py runserver"
createsuperuser = "python src/manage.py createsuperuser"
pre_migrate = "python src/manage.py makemigrations"
migrate = "python src/manage.py migrate"
shell = "python src/manage.py shell"
shellp = "python src/manage.py shell_plus"
test = "python src/manage.py test"
startapp = "python src/manage.py startapp {args}"
loaddata = "python src/manage.py loaddata {args}"
dumpdata = "python src/manage.py dumpdata {args}"
model = "python src/manage.py graph_models -a -g -o models.png"
req = "pdm export -f requirements --without-hashes -o requirements.txt"
check = "ruff check"
pre_format = "ruff check --fix"
format = "ruff format"